bug-hunt #260: quota - fix int32_t overflow in usage update functions
vfs_update_quota_blocks() and vfs_update_quota_inodes() cast
uint32_t cur_blocks/cur_inodes to int32_t before arithmetic.
When usage exceeds INT32_MAX, this cast produces a negative
value, causing the new_val < 0 clamp to reset usage to 0.
This gives the user effectively unlimited quota enforcement
and spuriously resets the soft-limit grace period.

Fix: use int64_t for the intermediate computation.

+++ src/fs/quota.c
@@ -254,7 +254,7 @@ void vfs_update_quota_blocks(uint16_t uid, int32_t delta_blocks)
     if (!eq)
         return;
 
-    int32_t new_val = (int32_t)eq->cur_blocks + delta_blocks;
+    int64_t new_val = (int64_t)eq->cur_blocks + delta_blocks;
     if (new_val < 0)
         new_val = 0;
 
@@ -275,7 +275,7 @@ void vfs_update_quota_inodes(uint16_t uid, int32_t delta)
     if (!eq)
         return;
 
-    int32_t new_val = (int32_t)eq->cur_inodes + delta;
+    int64_t new_val = (int64_t)eq->cur_inodes + delta;
     if (new_val < 0)
         new_val = 0;
